# api/api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Integracion",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
...
        "/api/ayuntamientos": {
            "get": {
                "tags": [
                    "Ayuntamientos"
                ],
                "summary": "Mostrar ayuntamientos",
                "operationId": "ceadadd73eb9c9bd76c6849ddf603d54",
                "responses": {
                    "200": {
                        "description": "Se ha completado la petición."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/ayuntamientos/": {
            "post": {
                "tags": [
                    "Ayuntamientos"
                ],
                "summary": "Crear ayuntamientos",
                "operationId": "06bf5f10dd506080149466e4e5002cd1",
                "responses": {
                    "200": {
                        "description": "Se ha completado la petición."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/ayuntamientos/:id": {
            "get": {
                "tags": [
                    "Ayuntamientos"
                ],
                "summary": "Muestra el ayuntamiento por ID",
                "operationId": "16f9fa9059eefa5bf8e05a0bd8c39868",
                "responses": {
                    "200": {
                        "description": "Se ha completado la petición."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Ayuntamientos"
                ],
                "summary": "Actualizar ayuntamientos",
                "operationId": "d6f1aad0f2b8291fcbc01a586d7c8143",
                "responses": {
                    "200": {
                        "description": "Se ha completado la petición."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Ayuntamientos"
                ],
                "summary": "Borrar ayuntamientos",
                "operationId": "b1b3c9b9d12459b062769c3367d74a78",
                "responses": {
                    "200": {
                        "description": "Se ha completado la petición."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "scheme": "bearer",
                "description": "Enter token in format (Bearer Token)",
                "name": "Authorization",
                "in": "header"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}